Moved toplevel bitmask.hpp into legacy/.
