+++ cpp/src/utilities/cudf_utils.h
@@ -4,7 +4,7 @@
 #include <utilities/error_utils.hpp>
 #include <cudf/cudf.h>
 #include <bitmask/legacy/bitmask_ops.hpp>
-#include <cudf/bitmask.hpp>
+#include <cudf/legacy/bitmask.hpp>
 
 #include <cuda_runtime_api.h>
